Extract CLI App into its own module
