+++ src/lib.rs
@@ -2,4 +2,4 @@ pub mod app;
 pub mod cli;
 pub mod execution;
 pub mod telemetry;
-pub mod ui;
+pub mod tui;
