+++ src/tui/tabs/logs.rs
